refactor(cli): repl hygiene batch 3 — CPR, registry split, tests
- extract CPR stdin helpers and StreamingConsole from loop.py
- split command_registry agents, model, and session_cmds into packages
- relocate orchestration unit tests under routing/tests/orchestration
- fix pytest import collision by omitting orchestration __init__.py
- update test monkeypatch targets for split registry modules
