+++ app/cli/interactive_shell/command_registry/agents/__init__.py
@@ -1,0 +1,7 @@
+"""Slash command: ``/fleet`` (registered local AI agent fleet view)."""
+
+from __future__ import annotations
+
+from app.cli.interactive_shell.command_registry.agents.core import COMMANDS
+
+__all__ = ["COMMANDS"]
