+++ app/cli/interactive_shell/command_registry/agents/core.py
@@ -9,22 +9,16 @@ from __future__ import annotations
 
 import math
 import os
-import time
 from collections import defaultdict
-from collections.abc import Callable
-from contextlib import nullcontext
 from pathlib import Path
 
-from prompt_toolkit.patch_stdout import patch_stdout
 from pydantic import ValidationError
 from rich.console import Console
-from rich.live import Live
 from rich.markup import escape
-from rich.text import Text
 from rich.tree import Tree
 
-from app.analytics.events import Event
-from app.analytics.provider import get_analytics
+from app.cli.interactive_shell.command_registry.agents.kill import _cmd_agents_kill
+from app.cli.interactive_shell.command_registry.agents.trace import _cmd_agents_trace
 from app.cli.interactive_shell.command_registry.types import SlashCommand
 from app.cli.interactive_shell.runtime import ReplSession
 from app.cli.interactive_shell.ui import (
@@ -51,9 +45,7 @@ from app.fleet_monitoring.conflicts import (
 )
 from app.fleet_monitoring.coordination import BranchClaims
 from app.fleet_monitoring.discovery import registered_and_discovered_agents
-from app.fleet_monitoring.lifecycle import TerminateResult, terminate
 from app.fleet_monitoring.registry import AgentRegistry
-from app.fleet_monitoring.tail import AttachSession, AttachUnsupported, attach
 
 _AGENTS_FIRST_ARGS: tuple[tuple[str, str], ...] = (
     ("budget", "view or edit per-agent hourly budgets"),
@@ -66,52 +58,6 @@ _AGENTS_FIRST_ARGS: tuple[tuple[str, str], ...] = (
     ("graph", "render the wait-on dependency graph as a tree"),
     ("wait", "mark <pid> as waiting on another pid: /fleet wait <pid> --on <other-pid>"),
 )
-
-_TRACE_REFRESH_PER_SECOND = 10
-# Match the throttle period to ``Live``'s refresh rate: under a 1k-line/sec
-# agent the reader thread can publish chunks faster than Rich actually
-# paints, and each ``live.update(Text.from_ansi(...))`` we make in
-# excess just creates a Renderable Rich will discard at the next paint.
-# Throttling the *call* to ``Live`` to one period bounds CPU under burst
-# writers without affecting how fast the screen updates.
-_TRACE_RENDER_PERIOD_S = 1.0 / _TRACE_REFRESH_PER_SECOND
-# Cap the on-screen render to the most recent slice of the 4 MiB buffer
-# so we don't reparse a 4 MiB string through Rich at 10 fps under burst
-# writers. A few screens of context is plenty for "what is the agent
-# doing right now"; the full tail is still in ``sess.buffer`` for any
-# future drill-down view.
-_TRACE_RENDER_TAIL_BYTES = 64 * 1024
-
-
-def _render_trace_snapshot(live: Live, sess: AttachSession) -> None:
-    """Decode the bounded snapshot with UTF-8 boundary safety for ``Live``.
-
-    ANSI sequences are interpreted (Rich); treat traced output like unfiltered ``kubectl logs``.
-    """
-    snapshot = _slice_to_utf8_boundary(sess.buffer.snapshot(), _TRACE_RENDER_TAIL_BYTES)
-    live.update(Text.from_ansi(snapshot.decode("utf-8", errors="replace")))
-
-
-def _slice_to_utf8_boundary(data: bytes, max_bytes: int) -> bytes:
-    """Return the suffix of ``data`` that fits in ``max_bytes`` and starts
-    on a UTF-8 codepoint boundary.
-
-    A plain ``data[-max_bytes:]`` can land mid-codepoint, which decodes to
-    a leading U+FFFD under ``errors="replace"`` — visible as a stray
-    replacement character at the top of the live view. ``TailBuffer``
-    preserves boundaries by dropping whole chunks; we have to do the same
-    once we re-flatten and slice on the render side. UTF-8 continuation
-    bytes match ``10xxxxxx`` (``b & 0xC0 == 0x80``); a codepoint is at
-    most 4 bytes, so we walk forward at most 3 continuation bytes to
-    reach the next start byte.
-    """
-    if len(data) <= max_bytes:
-        return data
-    sliced = data[-max_bytes:]
-    start = 0
-    while start < 4 and start < len(sliced) and (sliced[start] & 0xC0) == 0x80:
-        start += 1
-    return sliced[start:]
 
 
 def _opensre_agent_id() -> str:
@@ -336,196 +282,6 @@ def _cmd_agents_budget(session: ReplSession, console: Console, args: list[str]) 
     return True
 
 
-# Type alias for the optional confirmation callback (used for testing).
-_ConfirmFn = Callable[[str], str]
-
-
-def _cmd_agents_kill(
-    session: ReplSession,
-    console: Console,
-    args: list[str],
-    *,
-    confirm_fn: _ConfirmFn | None = None,
-) -> bool:
-    """Handle ``/fleet kill <pid> [--force]``.
-
-    Sends SIGTERM, waits up to 5 s, then escalates to SIGKILL.
-    Asks for confirmation unless ``--force`` is present.
-    Emits an ``agent_killed`` analytics event on success.
-    """
-    force = "--force" in args
-    positional = [a for a in args if a != "--force"]
-
-    if not positional:
-        console.print(f"[{ERROR}]usage:[/] /fleet kill <pid> [--force]")
-        session.mark_latest(ok=False, kind="slash")
-        return True
-
-    raw_pid = positional[0]
-    try:
-        pid = int(raw_pid)
-    except ValueError:
-        console.print(f"[{ERROR}]invalid pid:[/] {escape(raw_pid)} is not an integer")
-        session.mark_latest(ok=False, kind="slash")
-        return True
-
-    if pid == os.getpid():
-        console.print(f"[{ERROR}]refusing to kill the opensre process itself[/]")
-        session.mark_latest(ok=False, kind="slash")
-        return True
-
-    # Look up agent name from registry for friendlier output.
-    registry = AgentRegistry()
-    record = registry.get(pid)
-    label = f"{record.name} (pid {pid})" if record else f"pid {pid}"
-
-    if not force:
-        prompt_text = f"About to SIGTERM {label}. Confirm? [y/N] "
-        if confirm_fn is not None:
-            answer = confirm_fn(prompt_text)
-        else:
-            answer = console.input(prompt_text)
-        if answer.strip().lower() not in ("y", "yes"):
-            console.print(f"[{DIM}]aborted.[/]")
-            return True
-
-    try:
-        result: TerminateResult = terminate(pid)
-    except ProcessLookupError:
-        console.print(f"[{ERROR}]no such process:[/] pid {pid}")
-        session.mark_latest(ok=False, kind="slash")
-        return True
-    except PermissionError:
-        console.print(f"[{ERROR}]permission denied:[/] cannot signal pid {pid}")
-        session.mark_latest(ok=False, kind="slash")
-        return True
-
-    if result.exited:
-        console.print(
-            f"[{HIGHLIGHT}]Sent {result.signal_sent}. "
-            f"Process exited after {result.elapsed_seconds:.1f}s.[/]"
-        )
-    else:
-        console.print(
-            f"[{WARNING}]Sent {result.signal_sent} but process may still be running "
-            f"after {result.elapsed_seconds:.1f}s.[/]"
-        )
-        session.mark_latest(ok=False, kind="slash")
-
-    # Remove from the agent registry so `/fleet` no longer shows the dead PID.
-    # Only forget when the process actually exited — otherwise it stays visible
-    # for further monitoring or another kill attempt.
-    if record is not None and result.exited:
-        registry.forget(pid)
-
-    event = Event.AGENT_KILLED if result.exited else Event.AGENT_KILL_FAILED
-    get_analytics().capture(
-        event,
-        {
-            "pid": str(pid),
-            "agent_name": record.name if record else "unknown",
-            "signal": result.signal_sent,
-            "exited": result.exited,
-            "elapsed_seconds": str(round(result.elapsed_seconds, 2)),
-        },
-    )
-    return True
-
-
-def _render_live_tail(console: Console, label: str, sess: AttachSession) -> None:
-    """kubectl-logs-style render: a single Ctrl+C returns to the prompt.
-
-    Catches :class:`KeyboardInterrupt` inside the ``Live`` block and
-    swallows it so the REPL doesn't see a traceback. ``stream_to_console``
-    in ``streaming.py`` uses a double-press pattern because it's
-    rendering an LLM response that the user might *not* want to abort
-    on a stray keypress; a logs-style view is the inverse — one press
-    is the canonical "stop" signal.
-    """
-    console.print(f"[{BOLD_BRAND}]trace {escape(label)}[/]  [{DIM}]Ctrl+C to stop[/]")
-    isatty = getattr(console.file, "isatty", None)
-    stdout_context = patch_stdout(raw=True) if callable(isatty) and isatty() else nullcontext()
-    try:
-        with (
-            stdout_context,
-            Live(
-                Text(""),
-                console=console,
-                refresh_per_second=_TRACE_REFRESH_PER_SECOND,
-                transient=False,
-                vertical_overflow="visible",
-            ) as live,
-        ):
-            # Iterating ``sess`` is what drains the reader queue and
-            # appends to ``sess.buffer`` — the loop body only needs the
-            # *side effect* of advancing, not the chunk value, so the
-            # iteration variable is intentionally discarded.
-            # Seeded at 0.0 so the first iteration always renders (any
-            # ``time.monotonic() - 0.0`` clears the period); the throttle
-            # kicks in from the second iteration onward.
-            last_render = 0.0
-            pending = False
-            for _ in sess:
-                now = time.monotonic()
-                if now - last_render >= _TRACE_RENDER_PERIOD_S:
-                    _render_trace_snapshot(live, sess)
-                    last_render = now
-                    pending = False
-                else:
-                    pending = True
-            # Final flush: the last chunk(s) may have arrived inside a
-            # throttle window; render once after the loop so the user
-            # sees the very latest state instead of whatever was on
-            # screen at the last gated update.
-            if pending:
-                _render_trace_snapshot(live, sess)
-    except KeyboardInterrupt:
-        # kubectl-logs-style: a single Ctrl+C ends the trace and returns
-        # to the REPL prompt without propagating a traceback. The
-        # ``with sess:`` in the caller still runs and joins the reader
-        # thread, so this swallow is safe.
-        pass
-    if sess.producer_exited:
-        # Distinguish "the agent died and we noticed" from "the user
-        # asked us to stop" so a long unattended trace doesn't look the
-        # same as a Ctrl+C abort.
-        console.print(f"[{DIM}]· process exited[/]")
-    console.print(f"[{DIM}]· trace ended[/]")
-
-
-def _cmd_agents_trace(session: ReplSession, console: Console, args: list[str]) -> bool:
-    """Live-tail an agent's stdout by pid; see :func:`_render_live_tail`.
-
-    Validates eagerly (``attach()`` raises :class:`AttachUnsupported`
-    synchronously on bad pid / unsupported fd type / missing file) so
-    we never enter the ``Live`` block on a target we cannot tail.
-    """
-    if len(args) != 1:
-        console.print(f"[{ERROR}]usage:[/] /fleet trace <pid>")
-        session.mark_latest(ok=False, kind="slash")
-        return True
-    try:
-        pid = int(args[0])
-    except ValueError:
-        console.print(f"[{ERROR}]invalid pid:[/] {escape(args[0])}")
-        session.mark_latest(ok=False, kind="slash")
-        return True
-
-    record = AgentRegistry().get(pid)
-    label = f"{record.name} (pid {pid})" if record else f"pid {pid}"
-
-    try:
-        sess = attach(pid)
-    except AttachUnsupported as exc:
-        console.print(f"[{ERROR}]cannot trace {escape(label)}:[/] {escape(exc.reason)}")
-        session.mark_latest(ok=False, kind="slash")
-        return True
-
-    with sess:
-        _render_live_tail(console, label, sess)
-    return True
-
-
 def _cmd_agents_wait(session: ReplSession, console: Console, args: list[str]) -> bool:
     """Handle ``/fleet wait <pid> --on <other-pid>``.
 
@@ -710,5 +466,3 @@ COMMANDS: list[SlashCommand] = [
         first_arg_completions=_AGENTS_FIRST_ARGS,
     ),
 ]
-
-__all__ = ["COMMANDS"]
