+++ mypy.ini
@@ -60,3 +60,8 @@ disable_error_code = import-untyped
 [mypy-app.tools.utils.data_validation]
 check_untyped_defs = False
 
+# Co-located routing orchestration tests (relocated from tests/cli/); keep mypy
+# scoped to app/ production code — these modules use pytest monkeypatch seams.
+[mypy-app.cli.interactive_shell.routing.tests.orchestration.*]
+ignore_errors = True
+
