--- tests/cli/interactive_shell/orchestration/__init__.py
@@ -1,1 +1,0 @@
-# Package marker for mirrored tests.
