+++ tests/fleet_monitoring/test_coordination.py
@@ -10,7 +10,7 @@ from types import SimpleNamespace
 import pytest
 from rich.console import Console
 
-import app.cli.interactive_shell.command_registry.agents as agents_shell
+import app.cli.interactive_shell.command_registry.agents.core as agents_shell
 from app.fleet_monitoring.coordination import BranchClaim, BranchClaims
 from app.fleet_monitoring.registry import AgentRecord, AgentRegistry
